requirements/production.txt
Django >= 1.8.1,<1.9
kobo>=0.4.3
# requires: openldap-devel
python-ldap
# django rest framework
djangorestframework >= 3.2,<3.3
Markdown
django-filter==0.9.2
django-mptt>=0.7.1
django-cors-headers
djangorestframework-composed-permissions
